github codes/SohaibUsamaTestWork/requirements.txt
fastapi
uvicorn[satndard]
mysql-connector-python
pydantic